requirements.txt
fastmcp
python-dotenv
starlette
mcp
requests
pandas
openai
langchain
langchain-community
langchain-core
langchain-openai
langchain-deepseek
pydantic
flask
ragas